UIDVALIDITY should be 64 bit according to the RFC3501.
svn path=/trunk/KDE/kdepimlibs/; revision=950063

+++ selectjob.h
@@ -54,7 +54,7 @@ class KIMAP_EXPORT SelectJob : public Job
     int recentCount() const;
     int firstUnseenIndex() const;
 
-    int uidValidity() const;
+    qint64 uidValidity() const;
     int nextUid() const;
 
   protected:
